Реорганизация func_tank.*
