+++ server/entities/base_tank.cpp
@@ -12,155 +12,9 @@
 *   without written permission from Valve LLC.
 *
 ****/
-#include "extdll.h"
-#include "util.h"
-#include "cbase.h"
-#include "effects.h"
-#include "weapons.h"
-#include "explode.h"
-#include "monsters.h"
-#include "player.h"
-
-
-#define SF_TANK_ACTIVE		0x0001
-
-#define SF_TANK_LINEOFSIGHT		0x0010
-#define SF_TANK_CANCONTROL		0x0020	// g-cont. can't activate tank if set for now
-
-// g-cont. reserve 1 bit for me
-#define SF_TANK_MATCHTARGET		0x0080
-#define SF_TANK_SOUNDON		0x8000
-
-#define MAX_CONTROLLED_TANKS		48
-
-enum TANKBULLET
-{
-	TANK_BULLET_NONE = 0,
-	TANK_BULLET_9MM = 1,
-	TANK_BULLET_MP5 = 2,
-	TANK_BULLET_12MM = 3,
-	TANK_BULLET_OTHER = 4,	// just for difference between pacific and armed tanks
-};
-
-// Custom damage
-// env_laser (duration is 0.5 rate of fire)
-// rockets
-// explosion?
+#include "base_tank.h"
 
 class CFuncTankControls;
-class CBaseTank : public CBaseDelay
-{
-	DECLARE_CLASS( CBaseTank, CBaseDelay );
-public:
-	void	Spawn( void );
-	void	Precache( void );
-	void	KeyValue( KeyValueData *pkvd );
-	void	Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
-	void	Think( void );
-	void	TrackTarget( void );
-
-	void TryFire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-	virtual void Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-	virtual Vector UpdateTargetPosition( CBaseEntity *pTarget )
-	{
-		// kill the player aiming jitter for scripted tanks
-		if( m_bulletType == TANK_BULLET_NONE && pTarget->IsPlayer( ))
-			return pTarget->Center() * 0.25f + pTarget->EyePosition() * 0.75f;
-		return pTarget->BodyTarget( GetAbsOrigin() );
-	}
-
-	void StartRotSound( void );
-	void StopRotSound( void );
-
-	// Bmodels don't go across transitions
-	virtual int ObjectCaps( void ) { return (BaseClass :: ObjectCaps() & ~FCAP_ACROSS_TRANSITION) | FCAP_HOLD_ANGLES; }
-
-	inline BOOL IsActive( void ) { return (pev->spawnflags & SF_TANK_ACTIVE) ? TRUE : FALSE; }
-	inline void TankActivate( void ) { pev->spawnflags |= SF_TANK_ACTIVE; SetNextThink( 0.1f ); m_iState = STATE_ON; m_fireLast = 0; }
-	inline void TankDeactivate( void ) { pev->spawnflags &= ~SF_TANK_ACTIVE; m_fireLast = 0; m_iState = STATE_OFF; StopRotSound(); }
-	inline BOOL CanFire( void ) { return (gpGlobals->time - m_lastSightTime) < m_persist; }
-	BOOL InRange( float range );
-
-	// Acquire a target. pPlayer is a player in the PVS
-	CBaseEntity *FindTarget( void );
-
-	void TankTrace( const Vector &vecStart, const Vector &vecForward, const Vector &vecSpread, TraceResult &tr );
-
-	Vector BarrelPosition( void )
-	{
-		matrix4x4	entity = EntityToWorldTransform();
-		return entity.VectorTransform( m_barrelPos );
-	}
-
-	void AdjustAnglesForBarrel( Vector &angles, float distance );
-	int IRelationship( CBaseEntity* pTarget );
-
-	DECLARE_DATADESC();
-
-	BOOL StartControl( CBasePlayer* pController, CFuncTankControls* pControls );
-	void StopControl( CFuncTankControls* pControls );
-
-	CFuncTankControls	*m_pControls;	// tankcontrols is used as a go-between.
-
-	virtual BOOL IsTank( void ) { return TRUE; }
-
-protected:
-	float		m_flNextAttack;
-	
-	float		m_yawCenter;	// "Center" yaw
-	float		m_yawRate;	// Max turn rate to track targets
-	float		m_yawRange;	// Range of turning motion (one-sided: 30 is +/- 30 degress from center)
-					// Zero is full rotation
-	float		m_yawTolerance;	// Tolerance angle
-
-	float		m_pitchCenter;	// "Center" pitch
-	float		m_pitchRate;	// Max turn rate on pitch
-	float		m_pitchRange;	// Range of pitch motion as above
-	float		m_pitchTolerance;	// Tolerance angle
-
-	float		m_fireLast;	// Last time I fired
-	float		m_fireRate;	// How many rounds/second
-	float		m_lastSightTime;	// Last time I saw target
-	float		m_persist;	// Persistence of firing (how long do I shoot when I can't see)
-	float		m_minRange;	// Minimum range to aim/track
-	float		m_maxRange;	// Max range to aim/track
-
-	Vector		m_barrelPos;	// Length of the freakin barrel
-	float		m_spriteScale;	// Scale of any sprites we shoot
-	int		m_iszSpriteSmoke;
-	int		m_iszSpriteFlash;
-	TANKBULLET	m_bulletType;	// Bullet type
-	int		m_iBulletDamage;	// 0 means use Bullet type's default damage
-	
-	Vector		m_sightOrigin;	// Last sight of target
-	int		m_spread;		// firing spread
-	int		m_iszMaster;	// Master entity (game_team_master or multisource)
-	int		m_iszFireMaster;
-	string_t		m_iszFireTarget;
-
-	int		m_iTankClass;	// Behave As
-};
-
-class CFuncTankControls : public CBaseDelay
-{
-	DECLARE_CLASS( CFuncTankControls, CBaseDelay );
-public:
-	virtual int ObjectCaps( void ) { return (BaseClass::ObjectCaps() & ~FCAP_ACROSS_TRANSITION) | FCAP_IMPULSE_USE; }
-	void Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
-	void HandleTank( CBaseEntity *pActivator, CBaseEntity *m_pTank, BOOL activate );
-	void KeyValue( KeyValueData *pkvd );
-	void Spawn( void );
-
-	DECLARE_DATADESC();
-
-	BOOL OnControls( CBaseEntity *pTest );
-
-	CBasePlayer	*m_pController;
-	Vector		m_vecControllerUsePos; // where was the player standing when he used me?
-	int		m_iTankName[MAX_CONTROLLED_TANKS]; // list if indexes into global string array
-	int		m_cTanks; // the total number of targets in this manager's fire list.
-	BOOL		m_fVerifyTanks;
-};
 
 BEGIN_DATADESC( CBaseTank )
 	DEFINE_KEYFIELD( m_yawRate, FIELD_FLOAT, "yawrate" ),
@@ -191,17 +45,6 @@ BEGIN_DATADESC( CBaseTank )
 	DEFINE_FIELD( m_sightOrigin, FIELD_VECTOR ),
 	DEFINE_FIELD( m_barrelPos, FIELD_VECTOR ),
 END_DATADESC()
-
-#define MAX_FIRING_SPREADS		ARRAYSIZE( gTankSpread )
-
-static Vector gTankSpread[] =
-{
-	Vector( 0, 0, 0 ),		// perfect
-	Vector( 0.025, 0.025, 0.025 ),// small cone
-	Vector( 0.05, 0.05, 0.05 ),	// medium cone
-	Vector( 0.1, 0.1, 0.1 ),	// large cone
-	Vector( 0.25, 0.25, 0.25 ),	// extra-large cone
-};
 
 void CBaseTank :: Spawn( void )
 {
@@ -919,430 +762,4 @@ void CBaseTank :: StopRotSound( void )
 	if( FBitSet( pev->spawnflags, SF_TANK_SOUNDON ))
 		STOP_SOUND( edict(), CHAN_STATIC, STRING( pev->noise ));
 	ClearBits( pev->spawnflags, SF_TANK_SOUNDON );
-}
-
-class CFuncTankGun : public CBaseTank
-{
-	DECLARE_CLASS( CFuncTankGun, CBaseTank );
-public:
-	void Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-};
-
-LINK_ENTITY_TO_CLASS( func_tank, CFuncTankGun );
-
-void CFuncTankGun::Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker )
-{
-	int i;
-
-	if( m_fireLast != 0 )
-	{
-		// fireBullets needs gpGlobals->v_up, etc.
-		UTIL_MakeAimVectors( GetAbsAngles( ));
-
-		int bulletCount = (gpGlobals->time - m_fireLast) * m_fireRate;
-
-		if( bulletCount > 0 )
-		{
-			for( i = 0; i < bulletCount; i++ )
-			{
-				switch( m_bulletType )
-				{
-				case TANK_BULLET_9MM:
-					FireBullets( 1, barrelEnd, forward, gTankSpread[m_spread], 4096, BULLET_MONSTER_9MM, 1, m_iBulletDamage, pevAttacker );
-					break;
-				case TANK_BULLET_MP5:
-					FireBullets( 1, barrelEnd, forward, gTankSpread[m_spread], 4096, BULLET_MONSTER_MP5, 1, m_iBulletDamage, pevAttacker );
-					break;
-				case TANK_BULLET_12MM:
-					FireBullets( 1, barrelEnd, forward, gTankSpread[m_spread], 4096, BULLET_MONSTER_12MM, 1, m_iBulletDamage, pevAttacker );
-					break;
-				case TANK_BULLET_NONE:
-				default: break;
-				}
-			}
-
-			BaseClass::Fire( barrelEnd, forward, pevAttacker );
-		}
-	}
-	else
-	{
-		BaseClass::Fire( barrelEnd, forward, pevAttacker );
-	}
-}
-
-class CFuncTankLaser : public CBaseTank
-{
-	DECLARE_CLASS( CFuncTankLaser, CBaseTank );
-public:
-	void	Activate( void );
-	void	KeyValue( KeyValueData *pkvd );
-	void	Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-	void	Think( void );
-	CLaser	*GetLaser( void );
-
-	DECLARE_DATADESC();
-private:
-	CLaser	*m_pLaser;
-	float	m_laserTime;
-};
-
-LINK_ENTITY_TO_CLASS( func_tanklaser, CFuncTankLaser );
-
-BEGIN_DATADESC( CFuncTankLaser )
-	DEFINE_FIELD( m_pLaser, FIELD_CLASSPTR ),
-	DEFINE_FIELD( m_laserTime, FIELD_TIME ),
-END_DATADESC()
-
-void CFuncTankLaser :: Activate( void )
-{
-	if( !GetLaser( ) )
-	{
-		ALERT( at_error, "Laser tank with no env_laser!\n" );
-		UTIL_Remove( this );
-		return;
-	}
-	else
-	{
-		m_pLaser->TurnOff();
-	}
-
-	m_bulletType = TANK_BULLET_OTHER;
-
-	BaseClass::Activate();
-}
-
-void CFuncTankLaser :: KeyValue( KeyValueData *pkvd )
-{
-	if( FStrEq( pkvd->szKeyName, "laserentity" ))
-	{
-		pev->message = ALLOC_STRING( pkvd->szValue );
-		pkvd->fHandled = TRUE;
-	}
-	else
-	{
-		BaseClass::KeyValue( pkvd );
-	}
-}
-
-CLaser *CFuncTankLaser :: GetLaser( void )
-{
-	if( m_pLaser )
-		return m_pLaser;
-
-	CBaseEntity *pEntity;
-
-	pEntity = UTIL_FindEntityByTargetname( NULL, STRING( pev->message ));
-
-	while( pEntity )
-	{
-		// Found the laser
-		if( FClassnameIs( pEntity->pev, "env_laser" ))
-		{
-			m_pLaser = (CLaser *)pEntity;
-			break;
-		}
-		else
-		{
-			pEntity = UTIL_FindEntityByTargetname( pEntity, STRING( pev->message ));
-		}
-	}
-	return m_pLaser;
-}
-
-void CFuncTankLaser :: Think( void )
-{
-	if( m_pLaser && ( gpGlobals->time > m_laserTime ))
-		m_pLaser->TurnOff();
-
-	CBaseTank::Think();
-}
-
-void CFuncTankLaser :: Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker )
-{
-	TraceResult tr;
-
-	if( m_fireLast != 0 && GetLaser() )
-	{
-		// TankTrace needs gpGlobals->v_up, etc.
-		UTIL_MakeAimVectors( GetAbsAngles( ));
-
-		int bulletCount = (gpGlobals->time - m_fireLast) * m_fireRate;
-		if( bulletCount )
-		{
-			for( int i = 0; i < bulletCount; i++ )
-			{
-				m_pLaser->SetAbsOrigin( barrelEnd );
-				TankTrace( barrelEnd, forward, gTankSpread[m_spread], tr );
-				
-				m_laserTime = gpGlobals->time;
-				m_pLaser->TurnOn();
-				m_pLaser->pev->dmgtime = gpGlobals->time - 1.0f;
-				m_pLaser->FireAtPoint( barrelEnd, tr );
-				m_pLaser->DontThink();
-			}
-
-			BaseClass::Fire( barrelEnd, forward, pev );
-		}
-	}
-	else
-	{
-		BaseClass::Fire( barrelEnd, forward, pev );
-	}
-}
-
-class CFuncTankRocket : public CBaseTank
-{
-	DECLARE_CLASS( CFuncTankRocket, CBaseTank );
-public:
-	void Precache( void );
-	void Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-};
-
-LINK_ENTITY_TO_CLASS( func_tankrocket, CFuncTankRocket );
-
-void CFuncTankRocket :: Precache( void )
-{
-	UTIL_PrecacheOther( "rpg_rocket" );
-	BaseClass::Precache();
-
-	m_bulletType = TANK_BULLET_OTHER;
-}
-
-void CFuncTankRocket :: Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker )
-{
-	if( m_fireLast != 0 )
-	{
-		int bulletCount = (gpGlobals->time - m_fireLast) * m_fireRate;
-
-		if( bulletCount > 0 )
-		{
-			for( int i = 0; i < bulletCount; i++ )
-			{
-				CBaseEntity *pRocket = CBaseEntity::Create( "rpg_rocket", barrelEnd, GetAbsAngles(), edict() );
-			}
-			BaseClass::Fire( barrelEnd, forward, pev );
-		}
-	}
-	else
-	{
-		BaseClass::Fire( barrelEnd, forward, pev );
-	}
-}
-
-class CFuncTankMortar : public CBaseTank
-{
-	DECLARE_CLASS( CFuncTankMortar, CBaseTank );
-public:
-	void KeyValue( KeyValueData *pkvd );
-	void Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker );
-	void Precache( void );
-};
-
-LINK_ENTITY_TO_CLASS( func_tankmortar, CFuncTankMortar );
-
-void CFuncTankMortar :: KeyValue( KeyValueData *pkvd )
-{
-	if( FStrEq( pkvd->szKeyName, "iMagnitude" ))
-	{
-		pev->impulse = Q_atoi( pkvd->szValue );
-		pkvd->fHandled = TRUE;
-	}
-	else
-	{
-		BaseClass::KeyValue( pkvd );
-	}
-}
-
-void CFuncTankMortar :: Precache( void )
-{
-	BaseClass::Precache();
-
-	m_bulletType = TANK_BULLET_OTHER;
-}
-
-void CFuncTankMortar::Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker )
-{
-	if( m_fireLast != 0 )
-	{
-		int bulletCount = (gpGlobals->time - m_fireLast) * m_fireRate;
-
-		// Only create 1 explosion
-		if( bulletCount > 0 )
-		{
-			TraceResult tr;
-
-			// TankTrace needs gpGlobals->v_up, etc.
-			UTIL_MakeAimVectors( GetAbsAngles());
-			TankTrace( barrelEnd, forward, gTankSpread[m_spread], tr );
-			ExplosionCreate( tr.vecEndPos, GetAbsAngles(), edict(), pev->impulse, TRUE );
-			BaseClass::Fire( barrelEnd, forward, pev );
-		}
-	}
-	else
-	{
-		BaseClass::Fire( barrelEnd, forward, pev );
-	}
-}
-
-//============================================================================
-// FUNC TANK CONTROLS
-//============================================================================
-LINK_ENTITY_TO_CLASS( func_tankcontrols, CFuncTankControls );
-
-BEGIN_DATADESC( CFuncTankControls )
-	DEFINE_FIELD( m_pController, FIELD_CLASSPTR ),
-	DEFINE_FIELD( m_vecControllerUsePos, FIELD_VECTOR ),
-	DEFINE_AUTO_ARRAY( m_iTankName, FIELD_STRING ),
-	DEFINE_FIELD( m_cTanks, FIELD_INTEGER ),
-	DEFINE_FIELD( m_fVerifyTanks, FIELD_BOOLEAN ),
-END_DATADESC()
-
-void CFuncTankControls :: Spawn( void )
-{
-	pev->solid = SOLID_TRIGGER;
-	pev->movetype = MOVETYPE_NONE;
-	pev->effects |= EF_NODRAW;
-
-	SET_MODEL( edict(), GetModel() );
-	UTIL_SetSize( pev, pev->mins, pev->maxs );
-
-	// store pev->target for backward compatibility
-	if( m_cTanks < MAX_CONTROLLED_TANKS && !FStringNull( pev->target ))
-	{
-		m_iTankName[m_cTanks] = pev->target;
-		m_cTanks++;
-	}
-
-	SetLocalAngles( g_vecZero );
-	SetBits( m_iFlags, MF_TRIGGER );
-}
-
-void CFuncTankControls :: KeyValue( KeyValueData *pkvd )
-{
-	// get support for spirit field too
-	if( FStrEq( pkvd->szKeyName, "parent" ) || FStrEq( pkvd->szKeyName, "movewith" ))
-	{
-		m_iParent = ALLOC_STRING(pkvd->szValue);
-		pkvd->fHandled = TRUE;
-	}
-	else if( m_cTanks < MAX_CONTROLLED_TANKS ) 
-	{
-		// add this field to the target list
-		// this assumes that additional fields are targetnames and their values are delay values.
-		char tmp[128];
-		UTIL_StripToken( pkvd->szKeyName, tmp );
-		m_iTankName[m_cTanks] = ALLOC_STRING( tmp );
-		m_cTanks++;
-		pkvd->fHandled = TRUE;
-	}
-}
-
-BOOL CFuncTankControls :: OnControls( CBaseEntity *pTest )
-{
-	if( m_hParent != NULL && FClassnameIs( m_hParent->pev, "func_tracktrain" ))
-	{
-		// transform local controller pos through tankcontrols pos
-		Vector vecTransformedControlerUsePos = EntityToWorldTransform().VectorTransform( m_vecControllerUsePos );
-		if(( vecTransformedControlerUsePos - pTest->GetAbsOrigin() ).Length() < 30 )
-		{
-			return TRUE;
-		}
-	}
-	else if(( m_vecControllerUsePos - pTest->GetAbsOrigin() ).Length() < 30 )
-	{
-		return TRUE;
-	}
-
-	return FALSE;
-}
-
-void CFuncTankControls :: HandleTank ( CBaseEntity *pActivator, CBaseEntity *m_pTank, BOOL activate )
-{
-	// it's tank entity
-	if( m_pTank && m_pTank->IsTank( ))
-	{
-		if( activate )
-		{
-			if(((CBaseTank *)m_pTank)->StartControl((CBasePlayer *)pActivator, this ))
-			{
-				m_iState = STATE_ON; // we have active tank!
-			}
-		}
-		else
-		{
-			((CBaseTank *)m_pTank)->StopControl( this );
-		}
-	}
-}
-
-void CFuncTankControls :: Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value )
-{ 
-	if( !m_pController && useType != USE_OFF )
-	{
-		if( !pActivator || !( pActivator->IsPlayer( )))
-			return;
-
-		if( m_iState != STATE_OFF || ((CBasePlayer *)pActivator)->m_pTank != NULL )
-			return;
-
-		// find all specified tanks
-		for( int i = 0; i < m_cTanks; i++ )
-		{
-			CBaseEntity *tryTank = NULL;
-	
-			// find all tanks with current name
-			while( (tryTank = UTIL_FindEntityByTargetname( tryTank, STRING( m_iTankName[i] ))))
-			{
-				HandleTank( pActivator, tryTank, TRUE );
-			}			
-		}
-
-		if( m_iState == STATE_ON )
-		{
-			// we found at least one tank to use, so holster player's weapon
-			m_pController = (CBasePlayer *)pActivator;
-			m_pController->m_pTank = this;
-
-			if( m_pController->m_pActiveItem )
-			{
-				m_pController->m_pActiveItem->Holster();
-				m_pController->pev->weaponmodel = 0;
-				// viewmodel reset in tank
-			}
-
-			m_pController->m_iHideHUD |= HIDEHUD_WEAPONS;
-
-			// remember where the player's standing, so we can tell when he walks away
-			if( m_hParent != NULL && FClassnameIs( m_hParent->pev, "func_tracktrain" ))
-			{
-				// transform controller pos into local space because parent can be moving
-				m_vecControllerUsePos = m_pController->EntityToWorldTransform().VectorITransform( m_pController->GetAbsOrigin() );
-			}
-			else m_vecControllerUsePos = m_pController->GetAbsOrigin();
-		}
-	}
-	else if( m_pController && useType != USE_ON )
-	{
-		// find all specified tanks
-		for( int i = 0; i < m_cTanks; i++ )
-		{
-			CBaseEntity *tryTank = NULL;
-
-			// find all tanks with current name
-			while(( tryTank = UTIL_FindEntityByTargetname( tryTank, STRING( m_iTankName[i] ))))
-			{
-				HandleTank( pActivator, tryTank, FALSE );
-			}			
-		}
-
-		// bring back player's weapons
-		if( m_pController->m_pActiveItem )
-			m_pController->m_pActiveItem->Deploy();
-
-		m_pController->m_iHideHUD &= ~HIDEHUD_WEAPONS;
-		m_pController->m_pTank = NULL;				
-
-		m_pController = NULL;
-		m_iState = STATE_OFF;
-	}
 }
