+++ server/entities/func_tankmortar.cpp
@@ -1,0 +1,61 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+#include "func_tankmortar.h"
+
+LINK_ENTITY_TO_CLASS( func_tankmortar, CFuncTankMortar );
+
+void CFuncTankMortar :: KeyValue( KeyValueData *pkvd )
+{
+	if( FStrEq( pkvd->szKeyName, "iMagnitude" ))
+	{
+		pev->impulse = Q_atoi( pkvd->szValue );
+		pkvd->fHandled = TRUE;
+	}
+	else
+	{
+		BaseClass::KeyValue( pkvd );
+	}
+}
+
+void CFuncTankMortar :: Precache( void )
+{
+	BaseClass::Precache();
+
+	m_bulletType = TANK_BULLET_OTHER;
+}
+
+void CFuncTankMortar::Fire( const Vector &barrelEnd, const Vector &forward, entvars_t *pevAttacker )
+{
+	if( m_fireLast != 0 )
+	{
+		int bulletCount = (gpGlobals->time - m_fireLast) * m_fireRate;
+
+		// Only create 1 explosion
+		if( bulletCount > 0 )
+		{
+			TraceResult tr;
+
+			// TankTrace needs gpGlobals->v_up, etc.
+			UTIL_MakeAimVectors( GetAbsAngles());
+			TankTrace( barrelEnd, forward, gTankSpread[m_spread], tr );
+			ExplosionCreate( tr.vecEndPos, GetAbsAngles(), edict(), pev->impulse, TRUE );
+			BaseClass::Fire( barrelEnd, forward, pev );
+		}
+	}
+	else
+	{
+		BaseClass::Fire( barrelEnd, forward, pev );
+	}
+}
